Adding the first instantiation of the bot
Technically the second.  The first was getting commands working.  This
added !info as well as automatic track updates.

I also took out my personal script data so that anyone that stumbles
upon this can get a little bit of guidance on how to use it.

Added a lot of comments.